remove builder setter, add 2 empty test classes
