+++ base/src/main/java/com/smartdevicelink/managers/screen/BaseScreenManager.java
@@ -63,7 +63,6 @@ abstract class BaseScreenManager extends BaseSubManager {
 	private TextAndGraphicManager textAndGraphicManager;
 	private VoiceCommandManager voiceCommandManager;
 	private MenuManager menuManager;
-	private Boolean enableDynamicMenuUpdates;
 
 	// Sub manager listener
 	private final CompletionListener subManagerListener = new CompletionListener() {
@@ -91,10 +90,9 @@ abstract class BaseScreenManager extends BaseSubManager {
 		}
 	};
 
-	BaseScreenManager(@NonNull ISdl internalInterface, @NonNull FileManager fileManager, Boolean enableDynamicMenuUpdates) {
+	BaseScreenManager(@NonNull ISdl internalInterface, @NonNull FileManager fileManager) {
 		super(internalInterface);
 		this.fileManager = new WeakReference<>(fileManager);
-		this.enableDynamicMenuUpdates = enableDynamicMenuUpdates;
 		initialize();
 	}
 
@@ -111,7 +109,7 @@ abstract class BaseScreenManager extends BaseSubManager {
 		if (fileManager.get() != null) {
 			this.softButtonManager = new SoftButtonManager(internalInterface, fileManager.get());
 			this.textAndGraphicManager = new TextAndGraphicManager(internalInterface, fileManager.get(), softButtonManager);
-			this.menuManager = new MenuManager(internalInterface, fileManager.get(), enableDynamicMenuUpdates);
+			this.menuManager = new MenuManager(internalInterface, fileManager.get());
 		}
 		this.voiceCommandManager = new VoiceCommandManager(internalInterface);
 	}
